Bring new schmema system up to running v/m ingestion

+++ DB/structure/sql/samples/create_table_samples.sql
@@ -1,7 +1,7 @@
 create table samples (
 	id bigserial not null,
 	accession text not null,
-	sequence_id bigint not null,
+	sequence_id integer not null,
 
 	bio_project text,
 	bio_sample text,
